backend/requirements.txt
Flask==2.2.5
pymongo==4.4.0
dnspython==2.3.0
python-dotenv==1.0.0
requests==2.31.0
bcrypt==4.0.1
flask-cors==3.0.10
